fix for fzf
Some stty flags that were being set on both master and stlave ttys
were not needed, apparently. In particular, some combination of flags
on both prevented, for example, `fzf` and `vim` from capturing
`Enter` somehow.

Using `raw iutf8` for master and `sane` for slave ttys seems to have
sufficed to fix this and keep previous fetatures. I'm not sure if
`iutf8` is really needed on master, but it was not being set in `rpw
zsh` but my `zsh` usually has this enabled. Also, `sane` was not
needed for `zsh` but `bash` apparently inherits `-icanon` from parent
or something, which breaks the expected `cat` behavior. Using `sane`
fixed this for `bash` and seems to have no negative effect on `zsh.